[R2_MT9700_MTK] 1. 移除警告 [ALL] 2. 預設關閉 ENABLE_MSTV_UART_DEBUG

+++ TSUMR2/HK_R2/monitor_ap/CUSTOM/MTK/UI/INC/menudef.h
@@ -572,7 +572,7 @@ typedef struct
 
 typedef struct
 {
-    BYTE Flags;
+    WORD Flags;
     BYTE XPos, YPos;
     fpGetValue GetValue;
 } NumberType;
@@ -595,7 +595,7 @@ typedef struct
 #if Multi_Gauge // multi-gauge
 typedef struct
 {
-    BYTE Flags;
+    WORD Flags;
     BYTE XPos, YPos;
     fpGetValue GetValue;
 } GaugeType;
